Complicate the build process
I've set it up now so that the build will be done through gulp.
The new process for building this project from sources should be `npm install` followed by a `gulp docker` command.

This will mean that the dependencies used to test the application are the exact same ones that are packaged with the image.